caretransmission.c
'retData->triedCount' with type 'uint8_t' 8b(canonical type is 'unsigned char') is promoted to type 'int' 32b in 'milliTimeoutValue << retData->triedCount', then sign-extended to type 'unsigned long' 64b

+++ resource/csdk/connectivity/src/caretransmission.c
@@ -151,7 +151,7 @@ static bool CACheckTimeout(uint64_t currentTime, CARetransmissionData_t *retData
 #ifndef SINGLE_THREAD
     // #1. calculate timeout
     uint64_t milliTimeoutValue = retData->timeout / USECS_PER_MSEC;
-    uint64_t timeout = (milliTimeoutValue << retData->triedCount) * USECS_PER_MSEC;
+    uint64_t timeout = ((uint64_t)milliTimeoutValue << (uint64_t)retData->triedCount) * (uint64_t)USECS_PER_MSEC;
 
     if (currentTime >= retData->timeStamp + timeout)
     {
